docs: Switch to slightly shorter signature hash
Follows-up d9f0622 and 8b4d550, which added SHA-512 to the install
command which is so long that it requires a linebreak mid-hash
to ensure the README won't break the Gitiles page layout.

SHA-256 is probably still overkill for a < 4K file, but fits
without linebreaks or truncation, which is two fewer steps that
can go wrong during a Fresh release.

Change-Id: I98f35bed6d276f69d5734e3812ea30bd9c6426d4