fresh-node10: Add support for Podman
Podman is a daemonless tool that supports rootless container execution with
a close to identical CLI syntax as Docker.

If podman is installed and Docker isn't, it'll be used. It can also be
explicitly selected by passing the `-podman` option.

It doesn't support consistency bind options, so those are only added for
Docker users.

By default, the host UID/GID don't exist in the container (and therefore
can't write to the filesystem) unless we pass `--userns=keep-id` or use
root. But even getting root in the container doesn't give you any form
of root access on the host.

Bug: T259974
Change-Id: I4fedbf0518b8f794270543d4b4d3b196e1e9fd9d