test: Add test case for -env-sauce "none found" warning
Follows-up d51ff3c, which added a warning if no matching env
vars were found. This can help developers understand find the
source of the issue (e.g. forgotten to set a variable, not the
fault of Selenium or MW).

To test this, we need to start using `env -i` as otherwise my
own setting of SAUCE_ or MW_ makes the test fail (by succeeding
without the expected warning when the env vars aren't set in the
test case).

When using `env -i`, two other changes are needed:

1. BASE needs to be defined in a smarter way since the old
   readlink/dirname trick with "$0" no longer works in this mode.

2. PATH needs to explicitly include /usr/local/bin since we give
   that up at the same time we gave up MW_/SAUCE_ vars.

Change-Id: I40091c0ece44d6739ed25232a0ca714eceaf1ea9