fresh-node10: Emit warning if using -env and no MW found
Follows-up d51ff3c, which made it so that -env no longer causes
a silent failure if there are no matching env vars found (it previously
caused grep to exit the process, even if some matching vars where in
the env file rather than the bash env). It also introduced a warning
for -env-sauce.

This commit adds a similar warning for -env (aka -env-mw).

This is slightly less trivial than for d51ff3c, because we need
to gather vars from two places (bash env, and env file) and only
issue a warning if both empty.

Change-Id: I9c82a90233fc38ceab8195d30b1c38798ccaa34f