fresh-node10: Fix bug where containers from closed tabs didn't exit
There is an issue where if you close a terminal tab where you have a
temporary fresh-node container running, it would not stop the
'docker run' process, and thus Docker would not stop th
 temporary container or delete its data.

Reproducible with Docker on macOS, as well as with Podman on Fedora.

If you close your outer bash session, the fresh-node script invocation
did exit, so our trap function does run correctly, but
the 'docker run' process doesn't want to exit for some reason.

As a temporary workaround, give our container a custom name,
and explicitly call 'docker rm -f' from a trap function to delete it
if it still exists by that time. Note that 'docker run' ensures the
name we gave it is unique and not already in-use (if it is in use, it
will refuse to start the container).

Change-Id: I0f8e3d1683e61fade542007119165cb1d2ec6541