bin: Rename fresh-node10 to fresh-node
The installer already used this unversioned name. I had a vague plan
to support multiple versions in some future, which motivated me to
follow a versioned internal file name. But, if and when we do that,
e.g. to offer a version that's too new/experimental for people to use
by default, then we can just version-name that one instead.

For now, my recommendation for people wanting to use the current node10
images and the upcoming node12 side-by-side is to simply keep the
current one by renaming it locally. E.g. simply run:

```
mv /usr/local/bin/fresh-node{,10}
```

Change-Id: I9a0053fe8f699dd5e4b4aeafe4aa643fcae19890