test: Assert that the welcome message is up-to-date
To make this as stable and easy as possible, I have exposed a "dump
versions" command to the test via FRESH_INTERNAL_RUNCMD, which then
asserts that the dumped lines are contained in the welcome message.

As a convenient side-effect, this means one can now run this without
having to first edit the fresh-node script, thus removing one of the
three steps during the release process. Yay!

Change-Id: Id13e323cd52cab2197c9c5b34e9714fa69aa5211