fresh-install: Restore friendly message, prep for multiple bins
Follows-up Ib6a88cc437a. I forgot to put the success message in,
and also did not yet test what happens if the checksum or url fails.

Success:

> $ bin/fresh-install
> 🌱 Fresh is ready!

HTTP error:

> $ bin/fresh-install
> fresh-install: HTTP 404 at https://gerrit.wikimedia.org/g/fresh/+/21blabla/bin/fresh-node10?format=TEXT
> ❌

Bad checksum:

> $ bin/fresh-install
> fresh-install: checksum of fresh-node10 does not match
> ❌

Also refactor the local function in preparation for having multiple
bins to install, and restore the old installer for now so that
we can freely test fresh-instal without causing the documented install
command to install an unreleased version.

Change-Id: Ie4ff67d1d82e43304986e251db4ba4ac04b262aa