docs: Update outdated recommendation when needing sudo
In the last release, the install command changed so `sudo` now needs
to be put elsewhere in the command. This mainly applies to Linux,
but seems sensible to recommend on macOS as well, although on macOS
the permission error generally doesn't happen given how Homebrew
generally provisions that directory as non-root as a way to strike a
different security balance (some benefit, some downside).

Bug: T297286
Change-Id: Ia235f2942f0dc487d76bf8902b8ea9b624c9cf90