fresh-install: Support installing to ~/.local/bin for Linux
On Linux distributions `/usr/local/bin` is not writable by the user
unlike on macOS/Homebrew. The troubleshooting section recommends
`sudo` which is not ideal. This change adds a fallback to write to
the user home directory instead.

The XDG base directory specification defines `$HOME/.local/bin` for
user specific executable files:
https://specifications.freedesktop.org/basedir-spec/
https://www.freedesktop.org/software/systemd/man/file-hierarchy.html

When `/usr/local/bin` it not writable, check if `~/.local/bin` is in
the PATH. If it is, try creating it if not already, and then install
there instead.

Remove the obsolete troubleshooting section in README.md.

Example for typical Linux if this is your first home command:

  $ bin/install
  creating /Users/hashar/.local/bin
  installing fresh-node12 to /Users/hashar/.local/bin/fresh-node
  installing fresh-node14 to /Users/hashar/.local/bin/fresh-node14

Example for Linux when you already have $HOME/bin:

  $ bin/install
  installing fresh-node12 to /Users/hashar/bin/fresh-node
  installing fresh-node14 to /Users/hashar/bin/fresh-node14

Example hint that replaces the current permission error if nothing
is available (test with empty PATH):

  $ PATH= /usr/bin/python3 ./bin/fresh-install
  Tried the following places: /usr/local/bin, ~/.local/bin, ~/bin
  Set "PATH=$PATH:~/.local/bin" in ~/.bashrc and try again.
  fresh-install: Could not find writable installation folder

Bug: T282879
Change-Id: I95f18eb6f7c055840c82dd54e25f1406745540bd