fresh-node16,18,20: Hide "unrequested emulation" warning on Apple ARM
By default, given there are not yet ARM versions of these images,
`docker run` issues the following warning:

> WARNING: The requested image's platform does not match the detected
> host platform and no specific platform was requested.

Set a specific platform to acknowledge this. fresh-node generally
works on Apple ARM as-is for node, npm, jsduck and Firefox.

Chromium is not yet reliable under emulation, and fails to even start
depending on the Docker Desktop settings.

Change-Id: Ief379f815b30480fbc7cbfe18b4de30689391b54