Fix instance name to avoid spaces or other unsupported characters
This allow running fresh-node in folders with spaces in their name.

This is specifically about the current directory only, not about
parent directories.

Without this change, running fresh-node fails (at least when using
podman as backend, but probably for docker as well), as the
container names have to match (regex from podman):
`[a-zA-Z0-9][a-zA-Z0-9_.-]*`

While we don't have many repos or project folders which don't comply
with this, it could still be handy to fix.

Change-Id: Ifc270d3b1a1980ef88a80209294eff8db175db89