fresh-npm: write "command install is disabled" to stderr
All other errors were already written to sderr, but the one that is
most likely ("command is disabled") was missing this.

The end-result is that if you have some automated process run
'npm install' and report stderr on failure, there was no clue to
explain why it failed as stderr was a mere empty string. For example,
a Node.js script running
`cp.execSync('npm install', { encoding: 'utf8' });` would throw
an Error that says "Command failed: npm install" but no explaination
as to why.

With this patch, it will instead say
> Command failed: npm install
> fresh-npm ERR!  command 'install' is disabled or unknown.

Change-Id: Iecb6e49ba241bdca7c607240a52c091973d4250e